Change to swagger3 and mode handlers to own file.
