+++ requirements.txt
@@ -1,3 +1,4 @@
 matplotlib>=3.2.0
 aiohttp>=2.6.0
 aiohttp-swagger[performance]
+aiohttp-swagger3
